requirements.txt
Flask==2.2.5
cryptography==41.0.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
python-dotenv==1.0.0
Werkzeug==2.2.3
